+++ packages/web-platform/src/shared/decoders.ts
@@ -94,7 +94,8 @@ export const windowLayoutItemDecoder: Decoder<Glue42Workspaces.WindowLayoutItem>
         url: optional(nonEmptyStringDecoder),
         title: optional(string()),
         showCloseButton: optional(boolean()),
-        allowExtract: optional(boolean())
+        allowExtract: optional(boolean()),
+        isMaximized: optional(boolean())
     })
 });